Renaming methods to have better semantical names
