+++ examples/next/pages/index.tsx
@@ -1,10 +1,10 @@
-import { useDonationWidget } from '@everfund/react-sdk';
+import { useDonationCheckoutModal } from '@everfund/react-sdk';
 
 import * as React from 'react';
 
 const Home = () => {
   const [success, setSuccess] = React.useState(false);
-  const { isOpen, openDonationWidget } = useDonationWidget({
+  const { openModal } = useDonationCheckoutModal({
     code: 'j1q16MvJl5lqfji06bVFKei3HZkW',
     onSuccess: (data) => {
       console.log('Everfund success', data);
@@ -26,7 +26,7 @@ const Home = () => {
     <>
       <div className="app">
         <div className="buttonContainer">
-          <button className="button" onClick={() => openDonationWidget()}>
+          <button className="button" onClick={() => openModal()}>
             Donate Now
           </button>
         </div>
