+++ examples/react/vite.config.ts
@@ -3,6 +3,8 @@ import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import * as path from 'path';
 
+console.log(process.env.USE_SOURCE, process.env.USE_SOURCE === 'true')
+
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
