+++ examples/vanilla-js/src/main.ts
@@ -1,9 +1,9 @@
-import everfund from '@everfund/js-sdk';
+import { checkoutModal } from '@everfund/js-sdk';
 import '@everfund/example-css';
 
 document?.getElementById('donate-modal')?.addEventListener('click', () => {
   console.log('launch Everfund Modal');
-  everfund.donationWidget({
+  checkoutModal.modal({
     code: 'j1q16MvJl5lqfji06bVFKei3HZkW',
     onSuccess: function (onSuccessPayload) {
       console.log(onSuccessPayload);
