+++ package.json
@@ -3,6 +3,7 @@
   "packageManager": "pnpm@8.6.10",
   "scripts": {
     "build": "turbo run build",
+    "examples:build": "turbo run examples:build",
     "dev": "turbo run dev --concurrency=20",
     "format": "prettier --write \"**/*\"",
     "lint": "turbo run lint",
