+++ turbo.json
@@ -5,6 +5,9 @@
       "dependsOn": ["^build"],
       "outputs": ["dist/**"]
     },
+    "examples:build": {
+      "dependsOn": ["^build"]
+    },
     "lint": {},
     "test": {
       "dependsOn": ["^build"]
